+++ src/kernlib/include/stream.hpp
@@ -9,7 +9,7 @@ class Stream {
   virtual size_t read(void* dest, size_t size) = 0;
   virtual size_t size() = 0;
   virtual size_t seek(int64_t offset, char base = 0) = 0;
-  virtual Stream* substream(int64_t offset = -1, size_t limit = -1) = 0;
+  virtual Stream* substream(int64_t offset = -1, size_t limit = size_t(-1)) = 0;
   virtual bool eof() = 0;
   virtual char* readstr() = 0;
   virtual ~Stream() {}
@@ -26,7 +26,7 @@ class MemoryStream: public Stream {
   size_t read(void* dest, size_t size);
   size_t size() PURE;
   size_t seek(int64_t offset, char base = 0);
-  Stream* substream(int64_t offset = -1, size_t limit = -1);
+  Stream* substream(int64_t offset = -1, size_t limit = size_t(-1));
   char* readstr();
   bool eof() PURE;
 };
